backend/requirements.txt
python-dotenv
setuptools
typing-extensions
Flask-SQLAlchemy
flask-migrate
flask-bcrypt
flask-jwt-extended
playwright
gunicorn
gitpython
docx2pdf
PyPDF2>=3.0
psycopg2-binary
pillow 
pytesseract
flask
flask-cors
werkzeug
wheel
requests